# perf: Implement indexed method for `BitMapIter::nth`