fixes #22751

This PR introduces an indexed implementation for `nth()` on `BitmapIter` to avoid the observed O(n^2) behavior in case Null values are present on list operations.

A follow-up PR will change the set operation algorithm so that it re-uses the iterator rather recreating a new one repeatedly and calling `skip()` from the start every iteration, see e.g. https://github.com/pola-rs/polars/blob/de8ae363c3e4a22b5d6f93f95519b40817e2b087/crates/polars-ops/src/chunked_array/list/sets.rs#L133.

While each PR indepdently will fix the problem, it makes sense to merge both.

Before/after results on the issue MRE (using build-release on local dev machine, output filtered):

**Before**:
Multiplier: 1,000
Time for list_without_null: 0.010s
Time for list_with_null: 0.006s
Slowdown: 0.6x


Multiplier: 10,000
Time for list_without_null: 0.005s
Time for list_with_null: 0.467s
Slowdown: 92.3x


Multiplier: 100,000
Time for list_without_null: 0.023s
Time for list_with_null: 43.265s
Slowdown: 1853.1x


**After**:
Multiplier: 1,000
Time for list_without_null: 0.008s
Time for list_with_null: 0.001s
Slowdown: 0.1x


Multiplier: 10,000
Time for list_without_null: 0.003s
Time for list_with_null: 0.003s
Slowdown: 1.0x


Multiplier: 100,000
Time for list_without_null: 0.027s
Time for list_with_null: 0.028s
Slowdown: 1.0x